Allow for installing multiple versions of same pkg
We might need to install both mocked and non-mocked versions of a
package as part of the dependency chain from a single set of tests -
so we can't just use a map using the name as key to track what we have
done.

We now use processed as the base of our tracking, which uses the label
as a key - which is different for the two cases.  This then allows us
to install both versions of the package.

fixes #9.