Try and parse the mock output
Instead of discarding the mock output in our test code, instead write
it to a bytes.Buffer, and then try and parse it.  If the parse fails
then fail the test, but also write the mock output to a file to help
with debugging.  This should help catch cases where we generate broken
code ...