Start work on including real versions of code
The eventual aim is to be able to have runtime control of the
behaviour of our mocks.  Start by refactoring the code that processes
functions and methods so that we will have access to the original
code.  At the moment the original function body is not included - but
we have made the mockFile function nicer ...