Load function body
We now use the position information to read the function body text
from the source file.  We don't actually use it yet though, as we
aren't putting enough supporting code in for the generated code to
actually compile.

We need to not skip private functions & methods.  We also need to
include private struct fields.