Switch to runtime mocking
Instead of controlling mocking by choosing which version of a package
we import using rewriting, we now control mocking at runtime.  This
means that in general there is only one version of a package.  It also
means that we have the framework to make decisions about mocking on a
test by test basis.