Make literal/exprString into methods of mockGen
This means that exprString can have access to the information it needs
to recover the code from the original input file for func literals.
The previous approach of making func literals a special case was just
not going to work - as there are so many places a func literal can be
placed.