Implement loadInterfaceInfo
This is the last part of getting the interface mock generation
framework going.  We now have suport for embedded external interfaces.

We still need support for generating a package with mocks for
interfaces defined in the code under test.  However, that means
scoping types that are used, which we currently can't do.