Process imports from packages with non-Go code
When we process a package with non-Go code we have special handling,
and we weren't actually extracting the list of imports and adding it
to the list of imports to be processed.  We now handle this list of
imports as we do for any other package that we process.

fixes #18.