Add support for testing outside of GOPATH
If testing a package in the current directory, and it is outside of
the GOPATH, then withmock and mocktest would fail - as the calls to go
list to get various information would fail.  A couple of small tweaks,
mean that we can make this work - though an entire tree of packages
outside of GOPATH remains untested.

Hopefully this will fix issue #24.