Return imports from MakePkg
By returning an import map from MakePkg we not only say a call to
GetImports (and therefore a second parsing of the files in the
package), but we can also suppress packages that are not found when
the file in question contains build tags.

The assumption being that if a package is missing from a file with
build tags then it may well be that the file in question will not be
compiled and therefore that package isn't actually needed.  This
assumption may turn out to be false, but we need to wait 'til the go
tools try to build the package to find that out.

This was reported as issue #25.