Deal with stdlib packages using non-Go code
Many of the standard library packages use non-go code.  In this case
we don't actually care, since we only use generated code when mocking
the package.  So we can replace the external function/method with a
stub and still write a mock for it.  This replaces the special case we
were using for the os package.

Also, fix the issue #33 scenario now that it runs.

This fixes #33.