Start limiting stdlib to files for the platform
Instead of taking across and processing all files, we are now just
dealing with the files for the current GOOS/GOARCH combination.  This
work isn't complete as we only process file names - not the embedded
build constraints.  This is enough to make it possible to mock the net
package though.

Fixes #35