Add support for compilation mode
Add a flag that will allow the user to run mocktest in compilation
mode (i.e. call go test -c to just build the binary).  This means that
we can't fix the output for them (or rather we don't, as I can't be
bothered to go to the trouble of modifying the path strings in the
binary at the moment).

Also, we allow go test to tell the user that multiple packages aren't
supported currently - it might be nicer to do that check ourselves, as
then we won't waste time setting everything up.