Start refactoring for cache support
In order to be able to add support for caching mock generation we
first need to pull out the functionality that would be different when
mocked into one thing that we can replace.  This will be the
realPackage instance hiding behind the Package interface - then we can
use a cachePackage instance instead when pulling from the cache.