Add a new facility to provide mocks
In addition to automatically generated mocks indicated by // mock, we
can now use manually generated mocks by using // replace(import/path)
to use a different package in place of the named one.

This means that we can use manually generated mocks where we cannot
manage to generate an automatic mock (either due to bugs or to the way
the code is written).