Start working on better support for non-go code
This handles the case of non-go code where the code is .c or .s files
in the same package that are compiled by the gc compiler tools.

We don't currently respect the //export comments, which breaks cgo
code.

We also try to install extra packages that don't have any Go code in
them at all - which breaks things even more.