Rewrite injected functions in runtime
Since the runtime injects functions into other packages, when we
rename those functions, we also need to rename the one in runtime too.
We do this by supplying a rewriter to all stdlib packages to be filled
with the rewrite instructions from each package.  We then use this
rewriter to copy all the runtime files.