Work towards using our modified stdlib
Using our modified stdlib is not trivial - because now everything is
using the same stdlib.  Which means that when you enable mocking on a
package there needs to be some way for that to not affect test code
that shoud be using unmocked code.

The mechanism that I'm currently working on is a per-goroutine global
disable that will be injected into excluded packages.  With hooks to
inherit the state when a new goroutine is spawned.