Get rid of Package interface
Originally the idea was to have cached and non-chached implementations
of the Package interface.  I now think that it makes more sense to
have a single Package type, and have the cache interface be inside
that type - so that code using it is unaware that caching is being
performed.  The main thing is that we need to bring together all the
information about the package in order to know if a cached version
would be usable.