Working towards cache support
We can now store/retreive metadata in addition to file content in a
CacheFile.  This is important, as we want to be able to reproduce all
of the effects of processing a file, without having to process it.

Unfortunately this is probably going to be complicated by side effects
that are not immediately obvious (such as updating state in mockGen or
similar behaviour).