Enhance caching to support Install
We can now cache the results of building the packages (pkg.Install) -
which brings us closer to the end-goal of caching as much as possible.

This includes a lot of debug, partly to debug the changes - and partly
for timing investigations.  Having gotten things working I don't want
to risk breaking them, so I'm committing as is.