Fix init func naming
The way that we generated init function names wasn't good enough.  It
was fine when we always generated the entire package - but with
caching it becomes unreliable and can produce repeated function names.
Switch to a new mechanism that uses timestamps so that there is no
risk of creating duplicate function names.