Don't bother creating literal values for Mock_XXX types
The whole point of these types is so that you can create a mock object
that satifies a particular interface based on a private type of the
package.  The actual value isn't relevant, so we can just let Go use
the zero value and save ourselves the trouble of trying to figure out
the correct literal syntax.

Conflicts:
	lib/mock.go