Not sure that the new stdlib check was working
The check to detect that we had already processed a stdlib library may
not have been behaving correctly.  The main issue with processing the
same lib twice (other than the extra work) is the inability to create
the vendor link twice.  So now the code will not try to create the
link if it already exists (and exit with error if it exists but points
at the wrong place, or isn't a link).