Change the way we deal with internal packages
You can have internal packages that aren't inside the package that uses them -
which means that we don't install them into the custom gopath. So we need to
change the handling so that we will explictly install internal packages that
are needed.

This then means that we need to handle the issue we had previously where we
tried to install the same file twice. Previously we tried to avoid asking for
the files to be installed, now we will make the request, but notice the file is
already there.