Keep type behind build constraints
If we pick up a type from a file with build tags, then we should put the
related mock recorder into a file with the same tags. If we don't do this, then
they will be picked up at build time, and try to use the types that have been
hidden - causing the build to fail.