Start playing around with doing some code mods
Start trying to inject some extra code into the various files that we are
looking at.  The code we are adding right now isn't going to actually work, but
it's a start. We actually need to base the injected code on the function
signature so that it returns the correct types, and passes the parameters
through.