Don't include comments in the AST.
The problem with comments is that they are attached by position in the file,
which makes them very fragile when manipulating the AST. There is a CommentMap
that is supposed to help, but I'm not sure how to use it when adding code.
Disable comments for now, as they were causing some files to fail to be
formatted.

A better approach may be to use a DST (i.e. github.com/dave/dst) instead, so
that comments are attached to nodes instead of byte offsets. This should make
it easier to manipulate the code and maintain the comments in there correct
posistions.