Change the way that we handle methods
Originally I was trying to add a mock type into the various types so that we
can then manage the expectations within the instance. However, there are a
number of issues with this. The most basic of which being that adding the extra
field is an api breaking change and that it causes things to not compile. So,
change the way that we try to track methods by passing the receiver as an
argument to the single package-wide controller.