Start putting a i/f generator together
Create a new store that we can use to generate the interface file.

This is similar to what we do to generate the modified code, but in this case
we need to generate a different go.mod file, and we also need to generate new
source files rather than modifying the existing ones.