Start creating a mock generator too
This is similar to the interface generator, but the idea here is to generate
mocks for interfaces. So if a package provides any interfaces then this would
provide a packages that contains mocks that implements each of those
interfaces.

These mocks should also be useable in regular tests that aren't using the total
mocking approach.