Rework the way that the mock generator works
Split the mock generator into two parts. First we collect information about the
interface types that we want to create mocks for. Then we create the code in a
second pass. The reason for doing this is so that we can add a middle stage -
to correctly implement the mocks we need to resolve embedded interfaces so we
can implement the methods of those too.