Get basics of mocking stdlib working
You can now download a module of mocks for the standard library. Still need to
figure out exactly how the module names work, as we should be allowing
overlapping modules - e.g. io/fs should be a separate module to io, as it
should only be a sub-package in the same module.

Maybe the thing to do would have the standard library be a single mock module?