Use an error to get the mod file from the zip
By having a special error we can set it up so that the cache will extract the
mod file from the source zip if that error is returned, and do it in such a way
that both the source zip and the mod file are both added to the cache, so if
the source zip is wanted later we don't have to repeat the work.

This is much simpler than trying to do the extraction lower down, where we
can't then get the source zip into the cache without complications.